+++ packages/component-library/src/components/form/mt-checkbox/mt-checkbox.vue
@@ -29,7 +29,9 @@
           @inheritance-remove="$emit('inheritance-remove', $event)"
         >
           <template #label>
-            {{ label }}
+            <slot name="label">
+              {{ label }}
+            </slot>
           </template>
         </mt-base-field>
       </div>
@@ -440,6 +442,10 @@ export default defineComponent({
     &.checkbox-bordered {
       margin-bottom: var(--scale-size-8);
     }
+  }
+
+  & .mt-block-field__block {
+    min-height: unset;
   }
 }
 </style>