+++ examples/nuxt-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # nuxt-app
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [c6017dd]
+  - @shopware-ag/meteor-component-library@4.8.0
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [55b219b]
+  - @shopware-ag/meteor-component-library@4.7.1
+
 ## 0.0.27
 
 ### Patch Changes